+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   now):
   - [PyO3/pyo3-async-runtimes#60](https://github.com/PyO3/pyo3-async-runtimes/pull/60)
   - [PyO3/pyo3-async-runtimes#62](https://github.com/PyO3/pyo3-async-runtimes/pull/62)
+- `ryo3-uuid`
+  - supa speedy pydantic support added to `ry.uuid.UUID` (bc I need it now)
 - removed deprecated `.string()` methods from several structs/clases:
   - `ry.FsPath`
   - `ry.UUID`